created helper function to get options for txn index query
